perf(chat): decouple message conversion into a runtime-free module
convertToThreadMessageLike lived in GrueneratorChatRuntime.tsx alongside the
assistant-ui runtime, and the package barrel re-exported it. That static
re-export gave the eager graph an edge into the heavy runtime module. Move the
pure conversion (its only assistant-ui reference is an erased type) into its own
threadMessageConversion.ts and re-export from there, so the barrel no longer
pulls the runtime onto the initial load.
